+++ ST10440112-PROG6212-CCMS/ViewModels/ClaimSubmissionViewModel.cs
@@ -8,7 +8,7 @@ namespace ST10440112_PROG6212_CCMS.ViewModels
         [Required(ErrorMessage = "Hourly rate is required.")]
         [ValidHourlyRate(100, 1000)]
         [Display(Name = "Hourly Rate (R)")]
-        public int HourlyRate { get; set; }
+        public decimal HourlyRate { get; set; }
 
         [Required(ErrorMessage = "Total hours worked is required.")]
         [ValidHours(0.5f, 200f)]
@@ -16,7 +16,7 @@ namespace ST10440112_PROG6212_CCMS.ViewModels
         public float TotalHours { get; set; }
 
         [Display(Name = "Total Amount")]
-        public decimal TotalAmount => (decimal)(TotalHours * HourlyRate);
+        public decimal TotalAmount => (decimal)TotalHours * HourlyRate;
 
         [Display(Name = "Additional Notes")]
         [StringLength(1000, ErrorMessage = "Notes cannot exceed 1000 characters.")]
